Release 1.1.0 to develop

* Merged in feature/BRICR-801-project-journey (pull request #432)

[BRICR-801] implement project journey

* [BRICR-801] implement project journey

Approved-by: Aleksander Patschek

* Merged in fix/BRICR-929-int-to-fload (pull request #434)

[BRICR-929] change int to float in characteristic measurements

* [BRICR-929] change int to float in characteristic measurements

Approved-by: Michał Szuścik

* Merged in bugfix/BRICR-808-pim-media-pictures (pull request #433)

[BRICR-808] fix displaying pictures on pim media

* [BRICR-808] fix displaying pictures on pim media

* [BRICR-808] test fix

Approved-by: Aleksander Patschek

* Merged in bugfix/BRICR-000-project-list-item (pull request #435)

[BRICR-000] project list item fix

* [BRICR-000] project list item fix

Approved-by: Aleksander Patschek

* Merged in fix/BRICR-000-ncp-list (pull request #437)

Fix/BRICR-000 ncp list

* [BRICR-000] ncp list int to float

Approved-by: Michał Szuścik

* Merged in bugfix/BRICR-947-media-pictures-sort-caching (pull request #436)

[BRICR-947] remove caching sorting on media pictures

* [BRICR-947] remove caching sorting on media pictures

Approved-by: Aleksander Patschek

* [BRICR-000] update version

* Merged in bugfix/BRICR-000-edit-media-picture (pull request #439)

[BRICR-000] fix editing project media picture

* [BRICR-000] fix editing project media picture

* Merged in bugfix/BRICR-000-attention-note (pull request #440)

[BRICR-000] attention note in project list items

* [BRICR-000] attention note in project list items

* Merge branch 'develop' into release-1.1.0-to-develop

+++ src/app/projectDetails/sections/characteristics/CharacteristicsContainer.tsx
@@ -36,12 +36,12 @@ export const CharacteristicsContainer = () => {
             ...values,
             measurements: {
               ...values.measurements,
-              volumeFrom: parseInt(values?.measurements?.volumeFrom?.toString() ?? '') || null,
-              volumeTo: parseInt(values?.measurements?.volumeTo?.toString() ?? '') || null,
-              livingSpaceFrom: parseInt(values?.measurements?.livingSpaceFrom?.toString() ?? '') || null,
-              livingSpaceTo: parseInt(values?.measurements?.livingSpaceTo?.toString() ?? '') || null,
-              plotAreaFrom: parseInt(values?.measurements?.plotAreaFrom?.toString() ?? '') || null,
-              plotAreaTo: parseInt(values?.measurements?.plotAreaTo?.toString() ?? '') || null,
+              volumeFrom: parseFloat(values?.measurements?.volumeFrom?.toString() ?? '') || null,
+              volumeTo: parseFloat(values?.measurements?.volumeTo?.toString() ?? '') || null,
+              livingSpaceFrom: parseFloat(values?.measurements?.livingSpaceFrom?.toString() ?? '') || null,
+              livingSpaceTo: parseFloat(values?.measurements?.livingSpaceTo?.toString() ?? '') || null,
+              plotAreaFrom: parseFloat(values?.measurements?.plotAreaFrom?.toString() ?? '') || null,
+              plotAreaTo: parseFloat(values?.measurements?.plotAreaTo?.toString() ?? '') || null,
             },
             projectMarketing: {
               mainLogoId: values.projectMarketing?.mainLogoId,
